# Conflicts:
#	tetris-project/resource/cloud/tetris-resource/src/main/java/com/suma/venus/resource/task/BundleHeartBeatService.java

+++ tetris-project/resource/cloud/tetris-resource/src/main/java/com/suma/venus/resource/controller/BundleManageController.java
@@ -497,7 +497,7 @@ public class BundleManageController extends ControllerBase {
 //			}
 //		}
 
-		if (bundle.getDeviceIp() != null) {
+		if (bundle.getDeviceIp()!=null && "".equals(bundle.getDeviceIp())) {
 			bundleHeartBeatService.removeBundleStatus(bundle.getDeviceIp());
 		}
